requirements.txt
azure-ai-inference==1.0.0b7
requests==2.32.3
beautifulsoup4==4.12.3
matplotlib==3.9.2
numpy==2.1.2
ipywidgets==8.1.5
sentence-transformers==3.2.0
scikit-learn==1.5.2
python-dotenv==1.0.1
tf-keras==2.17.0